Solaris doesn't cast chars to ints in the is____() macros, e.g. isdigit(), isalpha() etc. => Added casts to the isdigit() and isalpha() casts
There doesn't seem to be a good way to use the TIFF
library to detect TILED TIFFs ...the behavior of
the libtiff library is different on different machines
so I have no choice but to skip the check for TILED
TIFFs ...darn, eh?  If someone throws a tiled TIFF
at the importer, it'll likely work but the output
may not be correct (I have no idea how to create
a tiled TIFF and don't know where to get one for
testing purposes ...they are not common.)

+++ src/libasf_import/arcgis_geotiff_support.c
@@ -16,6 +16,8 @@
 #include <assert.h>
 #include <ctype.h>
 #include <stdarg.h>
+#include <string.h>
+#include <strings.h>
 
 #include <geokeys.h>
 #include <geo_tiffp.h>
@@ -1181,11 +1183,11 @@ void Parse_ObjectString_to_Items (char objString[], ddItem *items, int *numItems
     item->indirectData = ' ';
 
     /* Parse out the number of items for this item type */
-    while (!isdigit(*pcTmp) && *pcTmp != '\0') pcTmp++;
+    while (!isdigit((int)*pcTmp) && *pcTmp != '\0') pcTmp++;
     pcTmp2 = pcTmp;
     numChars = 0;
     /*   ...Normally the following stops on ':' */
-    while (isdigit(*pcTmp2) && *pcTmp2 != '\0') {
+    while (isdigit((int)*pcTmp2) && *pcTmp2 != '\0') {
       numChars++;
       pcTmp2++;
     }
@@ -1216,11 +1218,11 @@ void Parse_ObjectString_to_Items (char objString[], ddItem *items, int *numItems
     {
       case _EMIF_T_ENUM:
         /* Parse out the number of items for this item type */
-        while (!isdigit(*pcTmp) && *pcTmp != '\0') pcTmp++;
+        while (!isdigit((int)*pcTmp) && *pcTmp != '\0') pcTmp++;
         pcTmp2 = pcTmp; /* pcTmp and pcTmp2 point at num of enums */
         numChars = 0;
         /*   ...Normally the following stops on ':' */
-        while (isdigit(*pcTmp2) && *pcTmp2 != '\0') {
+        while (isdigit((int)*pcTmp2) && *pcTmp2 != '\0') {
           numChars++;
           pcTmp2++;
         }
